refactor(engine): split engine_writer.go by concern
Splits the 1500-line god file into focused units along existing concern
boundaries. Zero behavior change — pure file moves.

- branch_mutations.go: lifecycle (create/delete/rename/checkout/merge/fetch/...)
- branch_tracking.go:  parent/scope/lock/freeze tracking
- commit_ops.go:       staging and commit primitives
- engine_worktree.go:  worktree registry + temp-worktree ops + WorktreeCheckoutMode/WorktreePruneMode
- pr_flags.go:         PR-body-update flagging
- stack_id.go:         stack-ID and stack-ref lifecycle (incl. GetStackIDsForBranches moved from remote_sync.go)

timeNow and sanitizeBranchNameForStackID now live in stack_id.go;
branchCheckedOutInAnotherWorktree lives in branch_mutations.go.

Phase 1 of the engine refactor runbook.
